--- src/components/model-item/tests/mockStore.js
@@ -1,5 +1,0 @@
-const mockStore = {
-  language: 0,
-  products: []
-};
-export default mockStore;
