--- src/components/model-item/tests/model-item.test.js
@@ -1,52 +1,0 @@
-import React from 'react';
-import Enzyme, { mount } from 'enzyme';
-import Adapter from 'enzyme-adapter-react-16';
-import * as reactRedux from 'react-redux';
-import { ThemeProvider } from '@material-ui/styles';
-
-import { BrowserRouter } from 'react-router-dom';
-
-import { theme } from '../../app/app-theme/app.theme';
-
-import mockStore from './mockStore';
-import ModelItem from '../model-item';
-
-Enzyme.configure({ adapter: new Adapter() });
-
-const themeValue = theme('light');
-let spyOnUseSelector;
-let spyOnUseDispatch;
-let mockDispatch;
-let wrapper;
-const name = [{}, { value: 'test' }];
-const model = { category: { name }, name, images: { small: '' } };
-
-describe('Model Item component tests', () => {
-  beforeEach(() => {
-    spyOnUseSelector = jest.spyOn(reactRedux, 'useSelector');
-    spyOnUseDispatch = jest.spyOn(reactRedux, 'useDispatch');
-
-    mockDispatch = jest.fn();
-
-    spyOnUseSelector.mockImplementation(() => mockStore);
-    spyOnUseDispatch.mockReturnValue(mockDispatch);
-
-    wrapper = mount(
-      <BrowserRouter>
-        <ThemeProvider theme={themeValue}>
-          <ModelItem model={model} />
-        </ThemeProvider>
-      </BrowserRouter>
-    );
-  });
-
-  afterEach(() => {
-    jest.restoreAllMocks();
-    spyOnUseSelector.mockClear();
-    wrapper = null;
-  });
-
-  it('Should render Model Item', () => {
-    expect(wrapper).toBeDefined();
-  });
-});
