--- src/containers/search-bar-list/search-bar-list-item/tests/mockStore.js
@@ -1,6 +1,0 @@
-const mockStore = {
-  language: 0,
-  currency: 0,
-  lightMode: true
-};
-export default mockStore;
