--- src/pages/home/models-list/tests/mockStore.js
@@ -1,5 +1,0 @@
-const mockStore = {
-  loading: false,
-  models: []
-};
-export default mockStore;
