--- src/pages/home/slider/tests/mockStore.js
@@ -1,9 +1,0 @@
-const testData = { 0: { value: 'test' } };
-
-const mockStore = {
-  language: 0,
-  images: {
-    items: [{ show: true, title: testData, description: testData, images: { large: 'test' } }]
-  }
-};
-export default mockStore;
