--- src/pages/home/slider/tests/slider.test.js
@@ -1,50 +1,0 @@
-import React from 'react';
-import Enzyme, { mount } from 'enzyme';
-import Adapter from 'enzyme-adapter-react-16';
-import * as reactRedux from 'react-redux';
-import { ThemeProvider } from '@material-ui/styles';
-
-import { BrowserRouter } from 'react-router-dom';
-
-import { theme } from '../../../../components/app/app-theme/app.theme';
-
-import mockStore from './mockStore';
-import SliderHomePage from '../slider-home-page';
-
-Enzyme.configure({ adapter: new Adapter() });
-
-const themeValue = theme('light');
-let spyOnUseSelector;
-let spyOnUseDispatch;
-let mockDispatch;
-let wrapper;
-
-describe('Slider component tests', () => {
-  beforeEach(() => {
-    spyOnUseSelector = jest.spyOn(reactRedux, 'useSelector');
-    spyOnUseDispatch = jest.spyOn(reactRedux, 'useDispatch');
-
-    mockDispatch = jest.fn();
-
-    spyOnUseSelector.mockImplementation(() => mockStore);
-    spyOnUseDispatch.mockReturnValue(mockDispatch);
-
-    wrapper = mount(
-      <BrowserRouter>
-        <ThemeProvider theme={themeValue}>
-          <SliderHomePage />
-        </ThemeProvider>
-      </BrowserRouter>
-    );
-  });
-
-  afterEach(() => {
-    jest.restoreAllMocks();
-    spyOnUseSelector.mockClear();
-    wrapper = null;
-  });
-
-  it('Should render Slider', () => {
-    expect(wrapper).toBeDefined();
-  });
-});
